esc_pos_printer/requirements.txt
Flask==2.2.2
python-escpos==3.0.0